Logger is now in the Concern module.
