+++ lib/concurrent/actor/behaviour/sets_results.rb
@@ -14,11 +14,11 @@ module Concurrent
           result = pass envelope
           if result != MESSAGE_PROCESSED && !envelope.future.nil?
             envelope.future.success result
-            log Logging::DEBUG, "finished processing of #{envelope.message.inspect}"
+            log DEBUG, "finished processing of #{envelope.message.inspect}"
           end
           nil
         rescue => error
-          log Logging::ERROR, error
+          log ERROR, error
           case error_strategy
           when :terminate!
             terminate!
