+++ lib/concurrent/actor/context.rb
@@ -1,3 +1,5 @@
+require 'concurrent/concern/logging'
+
 module Concurrent
   module Actor
 
@@ -22,6 +24,7 @@ module Concurrent
     class AbstractContext
       include TypeCheck
       include InternalDelegations
+      include Concern::Logging
 
       attr_reader :core
