+++ lib/concurrent/actor/default_dead_letter_handler.rb
@@ -2,7 +2,7 @@ module Concurrent
   module Actor
     class DefaultDeadLetterHandler < RestartingContext
       def on_message(dead_letter)
-        log Logging::INFO, "got dead letter #{dead_letter.inspect}"
+        log INFO, "got dead letter #{dead_letter.inspect}"
       end
     end
   end
